* Removed Bot-HandOff project since it has been relocated. * Verified the functionality of the new setup.
