--- Bot-HandOff/Bot-HandOff.nuspec
@@ -1,22 +1,0 @@
-<?xml version="1.0"?>
-<package >
-  <metadata>
-    <id>Bot-HandOff</id>
-    <version>0.0.1-alpha</version>
-    <title>Bot-HandOff</title>
-    <authors>Tomi Paananen, Jamie Dalton</authors>
-    <owners>Tomi Paananen, Jamie Dalton</owners>
-    <projectUrl>https://github.com/tompaana/intermediator-bot-sample</projectUrl>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>Unofficial Bot-HandOff library for bots built using the Microsoft Bot Framework</description>
-    <releaseNotes>Initial release</releaseNotes>
-    <copyright>MIT License</copyright>
-    <tags>bot framework human handoff handover agent supervisor microsoft c#</tags>
-    <language>en-US</language>
-  </metadata>
-
-  <files>
-    <file src="readme.md" target="" />
-    <file src="tools\Init.ps1" target="tools" />
-  </files>
-</package>
