--- Bot-HandOff/tools/Init.ps1
@@ -1,5 +1,0 @@
-﻿param($installPath, $toolsPath, $package, $project)
-
-$readme = $installPath + "\readme.md";
-
-$DTE.ExecuteCommand("File.OpenFile", $readme);
