+++ IntermediatorBotSample/Web.config
@@ -11,7 +11,7 @@
     <add key="MicrosoftAppPassword" value="" />
     <add key="owin:AutomaticAppStartup" value="false" />
 
-    <add key="PreventAddPendingRequestIfNoAggregationChannelExists" value="true" />
+    <add key="RejectPendingRequestIfNoAggregationChannel" value="true" />
 
     <!-- Add your connection string for routing data storage below -->
     <add key="RoutingDataStorageConnectionString" value="DefaultEndpointsProtocol=https;AccountName=account-name;AccountKey=account-key" />
